changes on the Animated class and two new methods addition
Additions:
animationsAll: object that will contain all the Animated instances.
createAnimated(): this new method will be called after a new instance of the Animated class is created. it is in charge of pushing to the animationsAll object the new Animated with its key equal to the id or className from the Animated.object
 getAll() method: This new method
returns an object that contains all animations instantiated in all the project(depending on where and when you call it)